Move Drawer to custom file
